+++ pkg/hub/api/api.go
@@ -1,9 +1,11 @@
 package api
 
 import (
+	"github.com/kobsio/kobs/pkg/hub/api/navigation"
 	"github.com/kobsio/kobs/pkg/hub/api/users"
 )
 
 type Config struct {
-	Users users.Config `json:"users"`
+	Users      users.Config      `json:"users"`
+	Navigation navigation.Config `json:"navigation"`
 }